requirements.txt
Django==2.1.7
djangorestframework==3.9.2
pandas==0.24.1
psycopg2==2.7.7
gunicorn==19.9.0
django-heroku==0.3.1
django-extensions==2.1.6
